pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beanie>=2.0.0",
    "fastapi>=0.118.0",
    "gunicorn>=23.0.0",
    "httpx>=0.28.1",
    "motor>=3.7.1",
    "pydantic-settings>=2.11.0",
    "uvicorn>=0.37.0",
    "pyjwt>=2.8.0",
    "google-auth>=2.29.0",
    "python-multipart>=0.0.9",
]
